databaseinterface/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
sqlalchemy
asyncpg